back/src/users/users.service.ts
import { Injectable } from '@nestjs/common';
import { User } from '../entity/Users';
import { Repository } from 'typeorm';
import { InjectRepository } from '@nestjs/typeorm';
import { CreateUserParams } from './types/types';

@Injectable()
export class UsersService {
  constructor(
    @InjectRepository(User) private userRepository: Repository<User>,
  ) {}
  createUser(userDetails: CreateUserParams) {
    const newUser = this.userRepository.create({ ...userDetails });
    return this.userRepository.save(newUser);
  }
  getAllUsers() {
    return this.userRepository.find();
  }
}
